Feat : feature/store_CRUD 20230913 1755

See merge request s09-fintech-finance-sub2/S09P22B207!39

+++ Front-End/pro/app/src/main/java/com/example/project/BarterPage.kt
@@ -97,8 +97,8 @@ fun BarterPage(navController: NavHostController) {
                     filter2Selected = "소분류"
                     viewModel.applyFilter(
                         BarterFilterDTO(
-                            filter1Selected,
-                            filter2Selected,
+                            filter1Selected.toInt(),
+                            filter2Selected.toInt(),
                             searchText,
                             currentPage
                         )
@@ -117,8 +117,8 @@ fun BarterPage(navController: NavHostController) {
                     filter2Selected = it
                     viewModel.applyFilter(
                         BarterFilterDTO(
-                            filter1Selected,
-                            filter2Selected,
+                            filter1Selected.toInt(),
+                            filter2Selected.toInt(),
                             searchText,
                             currentPage
                         )
